Moved scanner to build stage
